==> notebooks/2 - Machine Translation for TRANSLATE-TEST.ipynb <==
# import sys
# !{sys.executable} -m pip install transformers datasets transformers[torch] accelerate sentencepiece

# https://huggingface.co/spaces/evaluate-metric/sacrebleu
# https://huggingface.co/spaces/evaluate-metric/bertscore
# !{sys.executable} -m pip install evaluate sacrebleu bert_score

import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import evaluate

Translate the FR test set to English using the [NLLB-200](https://huggingface.co/facebook/nllb-200-distilled-600M) model:

! cd ../scripts && \
    python translate_fr2en.py --input-path "../data/mtob_domain_en_fr_test.csv" --output-path "../data/mtob_domain_en2fr_nllb_test.csv" --model-name "nllb"

Load the output generated by the above script:

test_df = pd.read_csv("../data/mtob_domain_en2fr_nllb_test.csv")

test_df.head(1)

Evaluate the translation quality using the [BLEU score](https://huggingface.co/spaces/evaluate-metric/sacrebleu):

sacrebleu = evaluate.load("sacrebleu")

translations = test_df["text_fr2en"].tolist()
references = [[t] for t in test_df["text_en"].values]

results = sacrebleu.compute(predictions=translations, references=references, lowercase=True)
print("BLEU score for FR -> EN:", round(results["score"], 1))

Evaluate the translation quality using [BERTScore](https://huggingface.co/spaces/evaluate-metric/bertscore):

bertscore = evaluate.load("bertscore")

translations = test_df["text_fr2en"].tolist()
references = test_df["text_en"].tolist()

results = bertscore.compute(predictions=translations, references=references, model_type="distilbert-base-uncased")
print("BERT score for FR -> EN:", round(np.mean(results["f1"]), 4))
