==> notebooks/4 - MLMs Finetuning & Inference.ipynb <==
import os
os.chdir("../scripts")

## A. Train on EN data 

### 1. Monolingual PLM (RoBERTa)

! python mlm_finetuning.py --dataset-dir "../data" \
    --train-filename "mtob_domain_en_fr_train.csv" --val-filename "mtob_domain_en_fr_val.csv" \
    --model-name "roberta" --language "english" --output-dir "../outputs/en_only" --run-name "roberta-1" \
    --model-max-length 50 --train-batch-size 128 --eval-batch-size 256 --learning-rate 0.00002 \
    --num-epochs 20 --eval-steps 100 --save-steps 100

# TRANSLATE-TEST
! python mlm_inference.py \
    --dataset-path "../data/mtob_domain_en2fr_nllb_test.csv" \
    --text-column "text_fr2en" \
    --model-path "../outputs/en_only/roberta-1/checkpoint-600/" \
    --output-path "../outputs/en_only/roberta-1/translate_test_preds.csv"

### 2. Multilingual PLM (XLM-RoBERTa)

! python mlm_finetuning.py --dataset-dir "../data" \
    --train-filename "mtob_domain_en_fr_train.csv" --val-filename "mtob_domain_en_fr_val.csv" \
    --model-name "xlm-roberta" --language "english" --output-dir "../outputs/en_only" --run-name "xlmr-en-1" \
    --model-max-length 50 --train-batch-size 128 --eval-batch-size 256 --learning-rate 0.00002 \
    --num-epochs 20 --eval-steps 100 --save-steps 100

# TRANSLATE-TEST
! python mlm_inference.py \
    --dataset-path "../data/mtob_domain_en2fr_nllb_test.csv" \
    --text-column "text_fr2en" \
    --model-path "../outputs/en_only/xlmr-en-1/checkpoint-1500/" \
    --output-path "../outputs/en_only/xlmr-en-1/translate_test_preds.csv"    
    
! python mlm_inference.py \
    --dataset-path "../data/mtob_domain_en2fr_nllb_test.csv" \
    --text-column "text_fr" \
    --model-path "../outputs/en_only/xlmr-en-1/checkpoint-1500/" \
    --output-path "../outputs/en_only/xlmr-en-1/test_preds.csv"    

## B. Train on FR data

### 1. Monolingual PLM (CamemBERT)

! python mlm_finetuning.py --dataset-dir "../data" \
    --train-filename "mtob_domain_en_fr_train.csv" --val-filename "mtob_domain_en_fr_val.csv" \
    --model-name "camembert" --language "french" --output-dir "../outputs/en_and_fr" --run-name "camembert-1" \
    --model-max-length 50 --train-batch-size 128 --eval-batch-size 256 --learning-rate 0.00002 \
    --num-epochs 20 --eval-steps 100 --save-steps 100

! python mlm_inference.py \
    --dataset-path "../data/mtob_domain_en2fr_nllb_test.csv" \
    --text-column "text_fr" \
    --model-path "../outputs/en_and_fr/camembert-1/checkpoint-800/" \
    --output-path "../outputs/en_and_fr/camembert-1/test_preds.csv"  

## 2. Multilingual PLM (XLM-RoBERTa)

! python mlm_finetuning.py --dataset-dir "../data" \
    --train-filename "mtob_domain_en_fr_train.csv" --val-filename "mtob_domain_en_fr_val.csv" \
    --model-name "xlm-roberta" --language "french" --output-dir "../outputs/en_and_fr" --run-name "xlmr-fr-1" \
    --model-max-length 50 --train-batch-size 128 --eval-batch-size 256 --learning-rate 0.00002 \
    --num-epochs 20 --eval-steps 100 --save-steps 100

! python mlm_inference.py \
    --dataset-path "../data/mtob_domain_en2fr_nllb_test.csv" \
    --text-column "text_fr" \
    --model-path "../outputs/en_and_fr/xlmr-fr-1/checkpoint-1200/" \
    --output-path "../outputs/en_and_fr/xlmr-fr-1/test_preds.csv"
