==> notebooks/5 - Zero-Shot LLM Prompting.ipynb <==
Prompt GPT-3.5 with instructions for a classification task.

# import sys
# !{sys.executable} -m pip install openai

import os
import warnings
warnings.filterwarnings("ignore")

import pandas as pd
from tqdm import tqdm

from openai import OpenAI
os.environ["OPENAI_API_KEY"] = ""

val_df = pd.read_csv("../data/mtob_domain_en_fr_val.csv")
test_df = pd.read_csv("../data/mtob_domain_en_fr_test.csv")

labels_fr = ["la messagerie", "l'appel", "l'événement", "la minuterie", "la musique", "le temps", "l'alarme", "les personnes", "le rappel", "les recettes", "les nouvelles"]
labels_en = ['messaging', 'calling', 'event', 'timer', 'music', 'weather', 'alarm', 'people', 'reminder', 'recipes', 'news']
labels_fr2en = dict(zip(labels_fr, labels_en))
label2id = {
    "alarm": 6,
    "calling": 1,
    "event": 2,
    "messaging": 0,
    "music": 4,
    "news": 10,
    "people": 7,
    "recipes": 9,
    "reminder": 8,
    "timer": 3,
    "weather": 5
}

en_prompt_template = (
    "Your task is to determine what category a phrase is related to. Both the phrase and the category will be in French.\n"
    "Possible categories: la messagerie, l'appel, l'événement, la minuterie, la musique, le temps, l'alarme, les personnes, le rappel, les recettes, les nouvelles.\n"
    "Phrase: {text_fr} \n"
    "Your answer (a single category from the list of possible categories): "
)
print(en_prompt_template)

model_name = "GPT-3.5"
model_version = "gpt-3.5-turbo"

client = OpenAI(
    base_url="https://openai.vocareum.com/v1",
    api_key=os.environ.get("OPENAI_API_KEY")
)

def query_openai(prompt, model_name=model_version):
    response = client.chat.completions.create(
        model=model_name,
        messages=[{
            "role": "user",
            "content": prompt
        }]
    )
    answer = response.choices[0].message.content
    return answer

example = val_df.query("label_text == 'news'").iloc[0]
example

en_prompt = en_prompt_template.format_map(example)
print(en_prompt)

query_openai(en_prompt)

output_dir = "../zero_shot_prompting"
os.makedirs(output_dir, exist_ok=True)

val_prompts = val_df.apply(lambda e: en_prompt_template.format_map(e), axis=1).tolist()
val_answers = [query_openai(p) for p in tqdm(val_prompts)]

val_preds_df = val_df[["id"]]
val_preds_df["label"] = [label2id.get(labels_fr2en.get(a, "unknown"), -1) for a in val_answers]

val_preds_path = os.path.join(output_dir, "val_preds.csv")
val_preds_df.to_csv(val_preds_path, index=False)
print(f"Saved val preds to {val_preds_path}")

test_prompts = test_df.apply(lambda e: en_prompt_template.format_map(e), axis=1).tolist()
test_answers = [query_openai(p) for p in tqdm(test_prompts)]

test_preds_df = test_df[["id"]]
test_preds_df["label"] = [label2id.get(labels_fr2en.get(a, "unknown"), -1) for a in test_answers]

test_preds_path = os.path.join(output_dir, "test_preds.csv")
test_preds_df.to_csv(test_preds_path, index=False)
print(f"Saved test preds to {test_preds_path}")
