==> notebooks/6 - Evaluation.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

from sklearn.metrics import f1_score, precision_recall_fscore_support

test_df = pd.read_csv("../data/mtob_domain_en2fr_nllb_test.csv")
test_df = test_df.rename({"label": "true_label"}, axis=1)
test_df.head(1)

preds_paths = {
    "base1_mostfreq": "../outputs/baselines/base1_mostfreq/test_preds.csv",
    "base2_stratified": "../outputs/baselines/base2_stratified/test_preds.csv",
    "base3_keywords": "../outputs/baselines/base3_keywords/test_preds.csv",
    "mlp_en_tt": "../outputs/en_only/mlp/translate_test_preds.csv",
    "roberta_en_tt": "../outputs/en_only/roberta-1/translate_test_preds.csv", 
    "xlm-roberta_en_tt": "../outputs/en_only/xlm-en-1/translate_test_preds.csv",
    "xlm-roberta_en": "../outputs/en_only/xlm-en-1/test_preds.csv",
    "camembert_fr": "../outputs/en_and_fr/camembert-1/test_preds.csv",
    "xlm-roberta_fr": "../outputs/en_and_fr/xlmr-fr-1/test_preds.csv",
    "gpt3.5_prompt": "../outputs/zero_shot_prompting/gpt-3.5/test_preds.csv"
}

preds_dfs = {}
for model, preds_path in preds_paths.items():
    preds_df = pd.read_csv(preds_path).rename({"label": "pred_label"}, axis=1)
    preds_df = test_df.merge(preds_df, left_on="id", right_on="id")
    preds_dfs[model] = preds_df.copy()

scores = []
for model, preds_df in preds_dfs.items():
    true_labels = preds_df["true_label"].values
    pred_labels = preds_df["pred_label"].values
    precision, recall, micro_f1, _ = precision_recall_fscore_support(
        true_labels, pred_labels, average="micro")
    macro_f1 = f1_score(true_labels, pred_labels, average="macro")
    weighted_f1 = f1_score(true_labels, pred_labels, average="weighted")
    scores.append([model, micro_f1, precision, recall, macro_f1, weighted_f1])
    
scores_df = pd.DataFrame(scores, columns=["Model", "F1", "Precision", "Recall", "Macro-F1", "Weighted-F1"])

scores_df

pdf = preds_dfs["gpt3.5_prompt"]
pdf[pdf.pred_label < 0].label_text.value_counts()
